#0000 Apply some basic and safe Java 7 optimisations from Intellij. Although a lot of files are touched, the changes are trivial and mostly just reduce clutter in the code.
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@24773 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ common/common-model/src/main/java/ish/oncourse/model/Student.java
@@ -61,7 +61,7 @@ public class Student extends _Student implements Queueable {
 
 	public String validateSchoolYear() {
 		if (getYearSchoolCompleted() != null) {
-			int givenYear = getYearSchoolCompleted().intValue();
+			int givenYear = getYearSchoolCompleted();
 			int thisYear = Calendar.getInstance().get(Calendar.YEAR);
 			if (givenYear > thisYear) {
 				return "Year school completed cannot be in the future if supplied.";
